move some search options to new company router
