+++ server/src/routes/index.ts
@@ -3,6 +3,7 @@ import exampleRouter from "./example";
 import authRouter from "./auth";
 import internshipModuleRouter from "./internshipModule";
 import internshipRouter from "./internship";
+import companyRouter from "./company";
 
 const router = Router();
 
@@ -13,5 +14,7 @@ router.use("/x", exampleRouter);
 router.use("/internship-modules", internshipModuleRouter);
 
 router.use("/internships", internshipRouter);
+
+router.use("/companies", companyRouter);
 
 export default router;
